6_03_extract_rules.patch
